allow email templates on action rules in component
